fixes for old GCC; removed thread local storage support

+++ src/engine/core/src/runtime/platform=pc/module-p=win32.cc
@@ -36,12 +36,12 @@ ref<const Module> Module::self()
         ::GetModuleInformation(process, hmodules[seen], &info, sizeof(info));
         if (seen == 0)
         {
-            s_module = ref<PE>::create(Arena::debug(), moduleName, (u64)info.lpBaseOfDll);
+            s_module = ref<PE>::create(Arena::debug(), moduleName, (u64)(uintptr_t)info.lpBaseOfDll);
             module = s_module;
         }
         else
         {
-            ref<Module> newModule = ref<PE>::create(Arena::debug(), moduleName, (u64)info.lpBaseOfDll);
+            ref<Module> newModule = ref<PE>::create(Arena::debug(), moduleName, (u64)(uintptr_t)info.lpBaseOfDll);
             module->m_next = newModule;
             module = newModule;
         }
